day 7 part 1 complete
wow, I can't believe that worked, honestly. This ended up working out as a base recursion, with a channel used to collect directories that were small enough to get included. I ran the base case in a goroutine and iterated over the hits channel in the main goroutine. This let me collect and return the size values.